lint and debt fixes to tests

+++ cellxgene_schema_cli/tests/fixtures/examples_validate.py
@@ -165,7 +165,7 @@ good_uns = {
 # ---
 # 4. Creating expression matrix,
 # X has integer values and non_raw_X has real values
-X = numpy.zeros([good_obs.shape[0], good_var.shape[0]])
+X = numpy.zeros([good_obs.shape[0], good_var.shape[0]], dtype=numpy.float32)
 non_raw_X = sparse.csr_matrix(X.copy())
 non_raw_X[0, 0] = 1.5
